requirements.txt
streamlit
openai
Pillow
numpy
opencv-python
requests
matplotlib
python-dotenv